+++ zigbee-api/src/main/java/org/bubblecloud/zigbee/network/zcl/ZclCommandMessage.java
@@ -35,7 +35,7 @@ public class ZclCommandMessage {
     /**
      * The source endpoint.
      */
-    private short sourceEnpoint;
+    private int sourceEnpoint;
     /**
      * The destination address.
      */
@@ -43,7 +43,7 @@ public class ZclCommandMessage {
     /**
      * The destination endpoint.
      */
-    private short destinationEndpoint;
+    private int destinationEndpoint;
     /**
      * The type.
      */
@@ -122,7 +122,7 @@ public class ZclCommandMessage {
      * Gets destination endpoint.
      * @return the destination endpoint
      */
-    public short getDestinationEndpoint() {
+    public int getDestinationEndpoint() {
         return destinationEndpoint;
     }
 
@@ -130,7 +130,7 @@ public class ZclCommandMessage {
      * Sets destination endpoint
      * @param destinationEndpoint the destination endpoint
      */
-    public void setDestinationEndpoint(final short destinationEndpoint) {
+    public void setDestinationEndpoint(final int destinationEndpoint) {
         this.destinationEndpoint = destinationEndpoint;
     }
 
@@ -170,7 +170,7 @@ public class ZclCommandMessage {
      * Gets source endpoint.
      * @return the source endpoint
      */
-    public short getSourceEnpoint() {
+    public int getSourceEnpoint() {
         return sourceEnpoint;
     }
 
@@ -178,7 +178,7 @@ public class ZclCommandMessage {
      * Sets source endpoint.
      * @param sourceEnpoint the source endpoint
      */
-    public void setSourceEnpoint(final short sourceEnpoint) {
+    public void setSourceEnpoint(final int sourceEnpoint) {
         this.sourceEnpoint = sourceEnpoint;
     }
